+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2018-06-05
+- Added FulfillerSelect component with support for calling Fulfiller Identity and Fulfillment Location services
+- Added language property (defaulting to 'eng') to all components
+
 ## [0.2.0] - 2018-05-16
 Simplified and renamed.